+++ fmpy/fmi1.py
@@ -77,6 +77,7 @@ callbacks.stepFinished = None
 
 
 class _FMU(object):
+    """ Base class for all FMUs """
 
     def __init__(self, guid, modelIdentifier, unzipDirectory, instanceName, logFMICalls=False):
 
@@ -84,7 +85,6 @@ class _FMU(object):
         self.modelIdentifier = modelIdentifier
         self.unzipDirectory = unzipDirectory
         self.instanceName = instanceName if instanceName is not None else self.modelIdentifier
-        self.fmuLocation = pathlib.Path(self.unzipDirectory).as_uri()
         self.logFMICalls = logFMICalls
 
         # remember the current working directory
@@ -319,9 +319,11 @@ class FMU1Slave(_FMU1):
     def instantiate(self, mimeType='application/x-fmu-sharedlibrary', timeout=0, visible=fmi1False,
                     interactive=fmi1False, functions=callbacks, loggingOn=fmi1False):
 
+        fmuLocation = pathlib.Path(self.unzipDirectory).as_uri()
+
         self.component = self.fmi1InstantiateSlave(self.instanceName.encode('UTF-8'),
                                                    self.guid.encode('UTF-8'),
-                                                   self.fmuLocation.encode('UTF-8'),
+                                                   fmuLocation.encode('UTF-8'),
                                                    mimeType.encode('UTF-8'),
                                                    timeout,
                                                    visible,