+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ## [Unreleased]
 
+### Fixed
+- resourceLocation in FMU2.instantiate() now points to resources directory instead of unzip directory
+
 ### Added
 - FMI functions (fmi1Reset, fmi2Reset, fmi2GetFMUState, fmi2SetFMUState, fmi2FreeFMUState)